Support two soroban host versions in preflight, switched by protocol.
