+++ Cargo.toml
@@ -8,15 +8,28 @@ members = [
 
 [workspace.package]
 rust-version = "1.74.0"
+[workspace.dependencies.soroban-env-host-prev]
+package = "soroban-env-host"
+version = "=21.2.0"
 
-[workspace.dependencies.soroban-env-host]
-version = "=21.1.0"
+[workspace.dependencies.soroban-env-host-curr]
+package = "soroban-env-host"
+version = "=22.0.0"
+git = "https://github.com/stellar/rs-soroban-env"
+rev = "b01e79899ceeeb3d39f9531bbc09b39bf0149da8"
 
-[workspace.dependencies.soroban-simulation]
-version = "=21.1.0"
+[workspace.dependencies.soroban-simulation-prev]
+package = "soroban-simulation"
+version = "=21.2.0"
+
+[workspace.dependencies.soroban-simulation-curr]
+package = "soroban-simulation"
+version = "=22.0.0"
+git = "https://github.com/stellar/rs-soroban-env"
+rev = "b01e79899ceeeb3d39f9531bbc09b39bf0149da8"
 
 [workspace.dependencies.stellar-xdr]
-version = "=21.1.0"
+version = "=21.2.0"
 features = [ "serde" ]
 
 [workspace.dependencies]
